refactor: separate usePartnerData
